refactor: change int to double for autoScrollingDelta

+++ packages/syncfusion_flutter_charts/lib/src/chart/axis/category_axis.dart
@@ -64,7 +64,7 @@ class CategoryAxis extends ChartAxis {
       RangeController? rangeController,
       double? maximumLabelWidth,
       double? labelsExtent,
-      int? autoScrollingDelta,
+      double? autoScrollingDelta,
       double? borderWidth,
       Color? borderColor,
       AxisBorderType? axisBorderType,
